+++ Gruntfile.js
@@ -30,11 +30,6 @@ module.exports = function( grunt ) {
 
 		// PHPLint
 		phplint: {
-			options: {
-				phpArgs: {
-					'-lf': null
-				}
-			},
 			all: [ 'src/**/*.php' ]
 		},