+++ src/BelfiusGateway.php
@@ -1,32 +1,32 @@
 <?php
 
 /**
- * Title: WooCommerce Maestro gateway
+ * Title: WooCommerce Belfius Direct Net gateway
  * Description:
  * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.2.2
- * @since 1.2.2
+ * @version 1.2.3
+ * @since 1.0.0
  */
-class Pronamic_WP_Pay_Extensions_WooCommerce_MaestroGateway extends Pronamic_WP_Pay_Extensions_WooCommerce_Gateway {
+class Pronamic_WP_Pay_Extensions_WooCommerce_BelfiusGateway extends Pronamic_WP_Pay_Extensions_WooCommerce_Gateway {
 	/**
 	 * The unique ID of this payment gateway
 	 *
 	 * @var string
 	 */
-	const ID = 'pronamic_pay_maestro';
+	const ID = 'pronamic_pay_belfius';
 
 	//////////////////////////////////////////////////
 
 	/**
-	 * Constructs and initialize an iDEAL gateway
+	 * Constructs and initialize an Belfius Direct Net gateway.
 	 */
 	public function __construct() {
 		$this->id             = self::ID;
-		$this->method_title   = __( 'Maestro', 'pronamic_ideal' );
-		$this->payment_method = Pronamic_WP_Pay_PaymentMethods::MAESTRO;
+		$this->method_title   = __( 'Belfius Direct Net', 'pronamic_ideal' );
+		$this->payment_method = Pronamic_WP_Pay_PaymentMethods::BELFIUS;
 
 		parent::__construct();
 	}
@@ -37,7 +37,7 @@ class Pronamic_WP_Pay_Extensions_WooCommerce_MaestroGateway extends Pronamic_WP_
 	function init_form_fields() {
 		parent::init_form_fields();
 
-		$this->form_fields['enabled']['label']       = __( 'Enable Maestro', 'pronamic_ideal' );
+		$this->form_fields['enabled']['label']       = __( 'Enable Belfius Direct Net', 'pronamic_ideal' );
 		$this->form_fields['description']['default'] = '';
 		$this->form_fields['icon']['default']        = '';
 	}