+++ src/Extension.php
@@ -3,11 +3,11 @@
 /**
  * Title: WooCommerce iDEAL Add-On
  * Description:
- * Copyright: Copyright (c) 2005 - 2016
+ * Copyright: Copyright (c) 2005 - 2017
  * Company: Pronamic
  *
  * @author Remco Tolsma
- * @version 1.2.1
+ * @version 1.2.3
  * @since 1.1.0
  */
 class Pronamic_WP_Pay_Extensions_WooCommerce_Extension {
@@ -41,6 +41,8 @@ class Pronamic_WP_Pay_Extensions_WooCommerce_Extension {
 			add_filter( 'pronamic_payment_redirect_url_' . self::SLUG, array( __CLASS__, 'redirect_url' ), 10, 2 );
 			add_action( 'pronamic_payment_status_update_' . self::SLUG, array( __CLASS__, 'status_update' ), 10, 1 );
 			add_filter( 'pronamic_payment_source_text_' . self::SLUG,   array( __CLASS__, 'source_text' ), 10, 2 );
+			add_filter( 'pronamic_payment_source_description_' . self::SLUG,   array( __CLASS__, 'source_description' ), 10, 2 );
+			add_filter( 'pronamic_payment_source_url_' . self::SLUG,   array( __CLASS__, 'source_url' ), 10, 2 );
 		}
 	}
 
@@ -74,6 +76,10 @@ class Pronamic_WP_Pay_Extensions_WooCommerce_Extension {
 
 		// @since 1.2.2
 		$gateways[] = 'Pronamic_WP_Pay_Extensions_WooCommerce_MaestroGateway';
+
+		// @since 1.2.3
+		$gateways[] = 'Pronamic_WP_Pay_Extensions_WooCommerce_BelfiusGateway';
+		$gateways[] = 'Pronamic_WP_Pay_Extensions_WooCommerce_KbcGateway';
 
 		return $gateways;
 	}
@@ -253,5 +259,23 @@ class Pronamic_WP_Pay_Extensions_WooCommerce_Extension {
 		);
 
 		return $text;
+	}
+
+	/**
+	 * Source description.
+	 */
+	public static function source_description( $description, Pronamic_Pay_Payment $payment ) {
+		$description = __( 'WooCommerce Order', 'pronamic_ideal' );
+
+		return $description;
+	}
+
+	/**
+	 * Source URL.
+	 */
+	public static function source_url( $url, Pronamic_Pay_Payment $payment ) {
+		$url = get_edit_post_link( $payment->source_id );
+
+		return $url;
 	}
 }